Solve result size mismatch
There is a problem of inconsistency in the result sizes between set and seq.
e.g., for AOL 10000, the exact result sizes are 10144 (seq) and 10131 (set).
However, the result size of seq has to be smaller or equal to that of set.
The problem is caused by duplicated outputs in seq.
By removing the duplications, the constraint has been satisfied.
Note that algorithms are still correct and unmodified.